requirements.txt
absl-py==0.11.0
aenum==3.0.0
affine==2.3.0
appnope==0.1.2
argon2-cffi==20.1.0
array2gif==1.0.4
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
boto3==1.17.60
botocore==1.20.60
branca==0.4.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colorful==0.5.4
cycler==0.10.0
dataclasses-json==0.5.3
decorator==4.4.2
defusedxml==0.6.0
earthengine-api==0.1.317
easydict==1.9
entrypoints==0.3
flatbuffers==1.12
folium==0.12.1
future==0.18.2
gast==0.3.3
GDAL==3.4.1
geetools==0.6.7
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.1
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
httplib2shim==0.0.3
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
matplotlib==3.3.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.6.1
nltk==3.2.4
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
osgeo==0.0.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
preprocessing==0.1.13
prettyprinter==0.18.0
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyproj==3.0.0.post1
pyrsistent==0.17.3
pyshp==2.1.3
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==21.0.1
qtconsole==5.0.1
QtPy==1.9.0
rasterio==1.1.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.4.2
scikit-image==0.18.2
scikit-learn==1.0.2
scipy==1.6.0
Send2Trash==1.5.0
sentinelhub==3.2.1
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
sphinx-rtd-theme==0.2.4
stringcase==1.2.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==3.1.0
tifffile==2021.4.8
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
utm==0.7.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
